Allowing for proper separation of Properties
In the initial implementation, you could not have a property field with the same name as another property field in a different property list. It would incorrectly assume that the field should be put in twice. Adding an id to each property, and prepending each property field's name with the id separated by a delimiter allows for proper separation between the property lists.